+++ src/subproviders/empty_wallet_subprovider.ts
@@ -1,4 +1,3 @@
-import * as Web3 from 'web3';
 import {JSONRPCPayload} from '../types';
 
 /*